# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Back2 Newsletters",
        "description": "Documentation Swagger pour la gestion des newsletters et archives.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur Local"
        }
    ],
    "paths": {
...
        "/newsletters": {
            "get": {
                "tags": [
                    "Newsletters"
                ],
                "summary": "Liste des newsletters (archives et projets)",
                "operationId": "897ed8f949066f311e7ede82d7de94b5",
                "responses": {
                    "200": {
                        "description": "Liste des newsletters récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Newsletter"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Newsletters"
                ],
                "summary": "Enregistrer une nouvelle newsletter (projet ou archive)",
                "operationId": "6c53b6e491b4bdfd3ce4b4ac87eb1a5b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sujet",
                                    "contenu",
                                    "statut"
                                ],
                                "properties": {
                                    "sujet": {
                                        "type": "string",
                                        "example": "Sujet de la newsletter"
                                    },
                                    "contenu": {
                                        "type": "string",
                                        "example": "Contenu HTML ou texte..."
                                    },
                                    "statut": {
                                        "type": "string",
                                        "example": "brouillon",
                                        "enum": [
                                            "brouillon",
                                            "envoyé"
                                        ]
                                    },
                                    "date_programmee": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "date_envoi": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Newsletter créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Newsletter"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/newsletters/{newsletter}": {
            "get": {
                "tags": [
                    "Newsletters"
                ],
                "summary": "Détails d'une newsletter",
                "operationId": "039d51aacd095ff3cdaeacfd943fa246",
                "parameters": [
                    {
                        "name": "newsletter",
                        "in": "path",
                        "description": "ID de la newsletter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Newsletter trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Newsletter"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Newsletter non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Newsletters"
                ],
                "summary": "Mettre à jour une newsletter",
                "operationId": "e59663f8d915021959b32fd753607c74",
                "parameters": [
                    {
                        "name": "newsletter",
                        "in": "path",
                        "description": "ID de la newsletter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sujet": {
                                        "type": "string"
                                    },
                                    "contenu": {
                                        "type": "string"
                                    },
                                    "statut": {
                                        "type": "string",
                                        "enum": [
                                            "brouillon",
                                            "envoyé"
                                        ]
                                    },
                                    "date_programmee": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "date_envoi": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Newsletter mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Newsletter"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Newsletters"
                ],
                "summary": "Supprimer une newsletter",
                "operationId": "287567068cecab58bb68d268752d7adf",
                "parameters": [
                    {
                        "name": "newsletter",
                        "in": "path",
                        "description": "ID de la newsletter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Newsletter supprimée avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/newsletters/{id}/send": {
            "post": {
                "tags": [
                    "Newsletters"
                ],
                "summary": "Envoyer une newsletter à tous les abonnés",
                "operationId": "4f6ee6574aad5fe9631337a883894e42",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la newsletter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sujet",
                                    "message"
                                ],
                                "properties": {
                                    "sujet": {
                                        "type": "string"
                                    },
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Newsletter envoyée avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Newsletter": {
                "title": "Newsletter",
                "description": "Objet représentant une newsletter (archive ou projet)",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "sujet": {
                        "type": "string",
                        "example": "Nouveau communiqué"
                    },
                    "contenu": {
                        "type": "string",
                        "example": "Contenu du communiqué..."
                    },
                    "statut": {
                        "type": "string",
                        "example": "brouillon",
                        "enum": [
                            "brouillon",
                            "envoyé"
                        ]
                    },
                    "date_programmee": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-03-25 10:00:00"
                    },
                    "date_envoi": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-03-19 15:30:00"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}